fixing the issue

+++ src/utils/cloudinary.js
@@ -20,7 +20,9 @@ const uploadOnCloudinary = async (localFilePath) => {
 
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary file as the upload operation got failed
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    } // remove the locally saved temporary file as the upload operation got failed
     return null;
   }
 };
